Working sensor drivers
vl53lox_Interrupt_Final is code that can be directly copied into the code running on the car's arduino. The "...debug" file can be used to see what's going on with the vl53lox_Interrupt_Final code if something is going wrong. It is functionally the same but it prints out the measurements when one or more have been updated, which the sensor interrupts make trickier than you might think, so I've provided code that does it.